Add files via upload
Initially I imported Train dataset and done Exploratory Data Analysis. In EDA we find missing values then plotted the bar plot of Missing values vs Non Missing values and here we conclude that as missing value increases 'SalePrice' increases in max. cases. After that we found the numbers of Numerical Features and then in numerical features we found temporal variables. and performed time series data analysis. In numerical variables we found numbers of Continuous Variables and Discrete Variables and analysed them. After that we find number of outliers. After that we found Categorical Features and then found numbers of Categorical Variables. In Feature Engineering we Replace Categorical Missing Variables in with 'Missing' and Numerical Missing Variables with median of that dataset because there was outliers in Features. This will be done with test and train datasets. After that we perform Time Series Analysis in Year Features. After with we handled Categorical Features in Train Datasets And Done Label Encoding in Categorical Variables. After that we have done Feature Scaling using MinMaxScaler in Test and Train Datasets. After that Feature Selection Will be Done using 'Lasso Regression' & 'SelectFromModel' After that we use HyperParameterTuning using RandomizedSearch using Certain Number of Features and then we Applied Random ForestRegressor and found RMSE.